:arrows_counterclockwise: Combine test utils into single module
